+++ NEWS.md
@@ -11,7 +11,7 @@ editor_options:
 -   Add missing metadata to spectra of Rosco gels.
 -   Add spectra for four Cooling (C), four Warming (W) and two "fluorescent" 
     (FL) colour correction photography filters from Hoya Tokina.
--   Rebuild data objects with 'photobiology' (0.12.0).
+-   Rebuild data objects with 'photobiology' (0.14.0).
 
 # photobiologyFilters 0.6.0